Fix C# code quality issues across src/
Agent-Logs-Url: https://github.com/KoshkiKode/cordite/sessions/db8bd75b-af19-4a07-9b72-51d1aedaef6a

+++ src/Core/DeterministicRng.cs
@@ -71,7 +71,7 @@ public class DeterministicRng
     public int NextInt(int max)
     {
         if (max <= 0) return 0;
-        ulong threshold = (ulong)(-(long)max % max);
+        ulong threshold = (ulong)(-(long)max) % (ulong)max;
         ulong r;
         do { r = NextUlong(); } while (r < threshold);
         return (int)(r % (ulong)max);
